test(currency): insere teste de integracao

+++ backend/src/usecases/remove-quote-to-currency.usecase.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { ICurrencyRepository } from '../domain/repository/ICurrencyRepository';
 
 @Injectable()
@@ -6,6 +6,15 @@ export class RemoveQuoteToCurrencyUsecase {
   constructor(private readonly currencyRepository: ICurrencyRepository) {}
 
   async exec(alias: string, quoteAlias: string) {
+    const getCurrency = await this.currencyRepository.findByAlias(alias);
+
+    if (!getCurrency) {
+      throw new BadRequestException('Currency not found');
+    }
+
+    if (!getCurrency.quotes.find((quote) => quote.alias === quoteAlias)) {
+      throw new BadRequestException('Currency not found');
+    }
     await this.currencyRepository.removeQuote(alias, quoteAlias);
   }
 }
